refactor: split run.rs into submodules (#133)
- run/mod.rs — RunSession, ExitDecision, check_docker, run_before_all
- run/env.rs — dotenv, GH token, extra-env tempfile, merge_env, token warning
- run/git.rs — resolve_git_identity, git_config_get
- run/summary.rs — write_last_run, parse_resume_state, resume_hint, exit_decision_from_summary

Tests move with their functions into the respective submodule #[cfg(test)] blocks.
No logic or public interface changes.
